backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
sqlalchemy>=2.0.35
psycopg[binary]>=3.2.2
pydantic-settings==2.3.4
alembic==1.13.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
